backend/requirements.txt
Flask
flask
flask-cors
google-generativeai
PyPDF2
python-dotenv